requirements.txt
altair==4.2.0
appnope==0.1.3
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
asgiref==3.4.1
asttokens==2.0.5
attrs==21.4.0
backcall==0.2.0
backports.entry-points-selectable==1.1.0
backports.zoneinfo==0.2.1
beautifulsoup4==4.11.1
bleach==5.0.0
blinker==1.4
cachetools==5.2.0
certifi==2021.5.30
cffi==1.14.6
charset-normalizer==2.0.6
click==8.1.3
commonmark==0.9.1
coreapi==2.3.3
coreschema==0.0.4
cryptography==3.4.8
cycler==0.11.0
debugpy==1.6.0
decorator==5.1.1
defusedxml==0.7.1
distlib==0.3.2
dj-database-url==0.5.0
Django==3.2.7
django-crispy-forms==1.12.0
django-ranged-response==0.2.0
django-simple-captcha==0.5.14
django-templated-mail==1.1.1
djangorestframework==3.12.4
djangorestframework-simplejwt==4.8.0
djoser==2.1.0
entrypoints==0.4
executing==0.8.3
fastjsonschema==2.15.3
filelock==3.0.12
fonttools==4.33.3
gitdb==4.0.9
GitPython==3.1.27
idna==3.2
importlib-metadata==4.11.4
importlib-resources==5.7.1
ipykernel==6.14.0
ipython==8.4.0
ipython-genutils==0.2.0
ipywidgets==7.7.0
itypes==1.2.0
jedi==0.18.1
Jinja2==3.0.1
joblib==1.1.0
jsonschema==4.6.0
jupyter-client==7.3.4
jupyter-core==4.10.0
jupyterlab-pygments==0.2.2
jupyterlab-widgets==1.1.0
kiwisolver==1.4.3
MarkupSafe==2.0.1
matplotlib==3.5.2
matplotlib-inline==0.1.3
mistune==0.8.4
nbclient==0.6.4
nbconvert==6.5.0
nbformat==5.4.0
nest-asyncio==1.5.5
notebook==6.4.12
numpy==1.22.4
oauthlib==3.1.1
packaging==21.3
pandas==1.4.2
pandocfilters==1.5.0
parso==0.8.3
patsy==0.5.2
pexpect==4.8.0
pickleshare==0.7.5
Pillow==8.3.2
platformdirs==2.3.0
prometheus-client==0.14.1
prompt-toolkit==3.0.29
protobuf==3.20.1
psutil==5.9.1
ptyprocess==0.7.0
pure-eval==0.2.2
pyarrow==8.0.0
pycparser==2.20
pydeck==0.7.1
Pygments==2.12.0
PyJWT==2.1.0
Pympler==1.0.1
pyparsing==3.0.9
pyrsistent==0.18.1
python-dateutil==2.8.2
python3-openid==3.2.0
pytz==2021.1
pytz-deprecation-shim==0.1.0.post0
pyzmq==23.1.0
requests==2.26.0
requests-oauthlib==1.3.0
rich==12.4.4
scikit-learn==1.1.1
scipy==1.8.1
seaborn==0.11.2
semver==2.13.0
Send2Trash==1.8.0
six @ file:///AppleInternal/BuildRoot/Library/Caches/com.apple.xbs/Sources/python3/python3-103/six-1.15.0-py2.py3-none-any.whl
sm==0.1.8
smmap==5.0.0
social-auth-app-django==4.0.0
social-auth-core==4.1.0
soupsieve==2.3.2.post1
sqlparse==0.4.2
stack-data==0.2.0
statsmodels==0.13.2
streamlit==1.10.0
tabulate==0.8.9
terminado==0.15.0
threadpoolctl==3.1.0
tinycss2==1.1.1
toml==0.10.2
toolz==0.11.2
tornado==6.1
traitlets==5.2.2.post1
typing_extensions==4.2.0
tzdata==2022.1
tzlocal==4.2
uritemplate==3.0.1
urllib3==1.26.6
validators==0.20.0
virtualenv==20.8.0
wcwidth==0.2.5
webencodings==0.5.1
widgetsnbextension==3.6.0
zipp==3.8.0
